requirements.txt
Pillow
loguru
pydantic
numpy